Created HandManager to split up some functionality from PlayerManager Moved associated tests from PlayerManagerTest to HandManagerTest Refactored calls to PlayerManager functions to HandManager functions
