+++ src/main/java/ScrabbleGame/ScrabbleGameController.java
@@ -22,6 +22,7 @@ public class ScrabbleGameController{
     
     private BoardManager boardManager;
     private PlayerManager playerManager;
+    private HandManager handManager;
     private GameLoaderSystem  gameLoader;
     private GameSaverSystem gameSaver;
     private ScoringSystem gameScorer;
@@ -36,6 +37,7 @@ public class ScrabbleGameController{
     public ScrabbleGameController(View v) {
         boardManager = new BoardManager(); // this class implements checkword checktile
         playerManager = new PlayerManager();// this class implements updatescore, drawtiles
+        handManager = new HandManager();
         gameLoader = new GameLoaderSystem();
         gameSaver = new GameSaverSystem();
         gameCreator = new GameCreator();
@@ -51,7 +53,7 @@ public class ScrabbleGameController{
         ArrayList<MoveInfo> moveInfos = boardManager.getMoves();
 
         for(MoveInfo move : moveInfos){
-            playerManager.addTile(game, move.getLetter());
+            handManager.addTile(game, move.getLetter());
             System.out.println("Letters");
             System.out.println(move.getLetter());
         }
@@ -60,9 +62,9 @@ public class ScrabbleGameController{
     }
     
     public void swapTiles() {
-        ((SwapHand) playerManager).swapHand(game);
+        ((SwapHand) handManager).swapHand(game);
         ((IncrementTurnUsecase) turnManager).incrementTurn(game);
-        ((FillHand)playerManager).fillHand(game);// fill the next player's hand
+        ((FillHand)handManager).fillHand(game);// fill the next player's hand
 
         view.updateView(game);
     }
@@ -70,7 +72,7 @@ public class ScrabbleGameController{
     public void placeTile(int[] coords, String letter) {
         boolean placeTileTrueness = ((PlaceTile) boardManager).checkLetter(coords, letter, game);
         if(placeTileTrueness){
-            ((RemoveTile)playerManager).removeTile(game, letter); // remove the tile
+            ((RemoveTile)handManager).removeTile(game, letter); // remove the tile
         }
         else{
             System.out.println("Invalid Move Played in placeTile");
@@ -99,13 +101,13 @@ public class ScrabbleGameController{
 
             ((IncrementTurnUsecase) turnManager).incrementTurn(game);
 
-            ((FillHand)playerManager).fillHand(game);// fill the next player's hand
+            ((FillHand)handManager).fillHand(game);// fill the next player's hand
         }
         else{
             ArrayList<MoveInfo> moves = boardManager.getMoves();
 
             for(MoveInfo move : moves){
-                playerManager.addTile(game, move.getLetter());
+                handManager.addTile(game, move.getLetter());
             }
         }
         boardManager.clearMoves();// reset moves for next turn
@@ -126,7 +128,7 @@ public class ScrabbleGameController{
     
     public void startGame(String[] names) { // create game usecase
         game = ((CreateGame)gameCreator).createNewGame(names);
-        ((FillHand)playerManager).fillHand(game);
+        ((FillHand)handManager).fillHand(game);
         view.updateView(game);
     }
